Update directory structure. Split files into relative directories
