+++ src/index.html
@@ -43,6 +43,5 @@
         <span class="rotate">//</span>
       </div>
     </footer>
-    <script src="index.js"></script>
   </body>
 </html>
